Connection
You can check if two fields are connected or not (you can set if the way
can go through enemy territory, and if the player's strongs can be used
as roads).
Notes:
My C++ code of object connection is also there (it's commented out).
Do not open Connection class, only if you really have to!